requirements.txt
backports.weakref==1.0.post1
brotlipy==0.7.0
charset-normalizer==3.1.0
eduvpn-client==4.1.1
eduvpn-common==1.1.0
et-xmlfile==1.1.0
jsonpointer==2.1
openpyxl==3.1.2
paho-mqtt==1.6.1
pandas==2.0.2
Pillow==9.4.0
ply==3.11
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
requests-auth==6.0.0
tzdata==2023.3
watchdog==2.3.0
webencodings==0.5.1
